fixed unsigned long problem between 64 bit and 32 bit

+++ N6LYT/ghpsdr3/trunk/src/server/receiver.h
@@ -42,7 +42,7 @@ typedef struct _receiver {
 RECEIVER receiver[MAX_RECEIVERS];
 
 typedef struct _buffer {
-    unsigned long sequence;
+    unsigned long long sequence;
     unsigned short offset;
     unsigned short length;
     unsigned char data[500];
